requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
loguru==0.7.2
packaging==24.2
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
